# core: integrate hmdb for big performance and integrity gains