--- a/clients/android/app/.gitignore
+++ b/clients/android/app/.gitignore
@@ -1,1 +1,2 @@
 /build
+/release