--- a/clients/android/app/build.gradle
+++ b/clients/android/app/build.gradle
@@ -11,8 +11,8 @@ android {
         applicationId "app.lockbook"
         minSdkVersion 22
         targetSdkVersion 32
-        versionCode 1
-        versionName "0.3.17"
+        versionCode 4
+        versionName "0.4.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }