--- a/clients/android/app/src/main/java/app/lockbook/model/CoreModel.kt
+++ b/clients/android/app/src/main/java/app/lockbook/model/CoreModel.kt
@@ -1,6 +1,5 @@
 package app.lockbook.model
 
-import app.lockbook.core.*
 import app.lockbook.util.*
 import com.github.michaelbull.result.Err
 import com.github.michaelbull.result.Result
@@ -17,9 +16,12 @@ import kotlinx.serialization.modules.subclass
 
 object CoreModel {
     private const val PROD_API_URL = "https://api.prod.lockbook.net"
-    fun getAPIURL(): String = System.getenv("API_URL") ?: PROD_API_URL
+    private fun getAPIURL(): String = System.getenv("API_URL") ?: PROD_API_URL
 
-    private fun <O, E : Enum<E>> SerializersModuleBuilder.createPolyRelation(okSerializer: KSerializer<O>, errSerializer: KSerializer<E>) {
+    private fun <O, E : Enum<E>> SerializersModuleBuilder.createPolyRelation(
+        okSerializer: KSerializer<O>,
+        errSerializer: KSerializer<E>
+    ) {
         polymorphic(IntermCoreResult::class) {
             subclass(IntermCoreResult.CoreOk.serializer(okSerializer))
             subclass(IntermCoreResult.CoreErr.serializer(errSerializer))
@@ -31,356 +33,280 @@ object CoreModel {
         }
     }
 
-    inline fun <reified C, reified E> Json.tryParse(json: String): Result<C, CoreError<E>>
-    where E : Enum<E>, E : UiCoreError = try {
+    private inline fun <reified C, reified E> Json.tryParse(json: String): Result<C, CoreError<E>>
+            where E : Enum<E>, E : UiCoreError = try {
         decodeFromString<IntermCoreResult<C, E>>(json).toResult()
     } catch (e: Exception) {
         Err(CoreError.Unexpected("Cannot parse json."))
     }
 
-    private val setUpInitLoggerParser = Json {
+    val setUpInitLoggerParser = Json {
         serializersModule = SerializersModule {
-            createPolyRelation(Unit.serializer(), InitLoggerError.serializer())
+            createPolyRelation(Unit.serializer(), InitError.serializer())
         }
     }
 
-    fun setUpInitLogger(path: String): Result<Unit, CoreError<InitLoggerError>> =
-        setUpInitLoggerParser.tryParse(initLogger(path))
+    fun init(config: Config): Result<Unit, CoreError<InitError>> =
+        setUpInitLoggerParser.tryParse(app.lockbook.core.init(setUpInitLoggerParser.encodeToString(config)))
 
-    val getDBStateParser = Json {
-        serializersModule = SerializersModule {
-            createPolyRelation(State.serializer(), GetStateError.serializer())
-        }
-    }
-
-    fun getDBState(config: Config): Result<State, CoreError<GetStateError>> =
-        getDBStateParser.tryParse(
-            getDBState(
-                getDBStateParser.encodeToString(config)
-            )
-        )
-
-    val migrateDBParser = Json {
-        serializersModule = SerializersModule {
-            createPolyRelation(Unit.serializer(), MigrationError.serializer())
-        }
-    }
-
-    fun migrateDB(config: Config): Result<Unit, CoreError<MigrationError>> =
-        migrateDBParser.tryParse(
-            migrateDB(
-                migrateDBParser.encodeToString(
-                    config
-                )
-            )
-        )
-
-    val createAccountParser = Json {
+    private val createAccountParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Account.serializer(), CreateAccountError.serializer())
         }
     }
 
     fun createAccount(
-        config: Config,
         account: String
     ): Result<Account, CoreError<CreateAccountError>> = createAccountParser.tryParse(
-        createAccount(
-            createAccountParser.encodeToString(config),
+        app.lockbook.core.createAccount(
             account,
             getAPIURL()
         )
     )
 
-    val importAccountParser = Json {
+    private val importAccountParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Account.serializer(), ImportError.serializer())
         }
     }
 
-    fun importAccount(config: Config, account: String): Result<Account, CoreError<ImportError>> = importAccountParser.tryParse(
-        importAccount(
-            importAccountParser.encodeToString(
-                config
-            ),
-            account
+    fun importAccount(account: String): Result<Account, CoreError<ImportError>> =
+        importAccountParser.tryParse(
+            app.lockbook.core.importAccount(
+                account
+            )
         )
-    )
 
-    val exportAccountParser = Json {
+    private val exportAccountParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(String.serializer(), AccountExportError.serializer())
         }
     }
 
-    fun exportAccount(config: Config): Result<String, CoreError<AccountExportError>> =
+    fun exportAccount(): Result<String, CoreError<AccountExportError>> =
         exportAccountParser.tryParse(
-            exportAccount(
-                exportAccountParser.encodeToString(config)
-            )
+            app.lockbook.core.exportAccount()
         )
 
-    val syncAllParser = Json {
+    private val syncAllParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Unit.serializer(), SyncAllError.serializer())
         }
     }
 
-    fun syncAll(config: Config, syncModel: SyncModel?): Result<Unit, CoreError<SyncAllError>> =
+    fun syncAll(syncModel: SyncModel?): Result<Unit, CoreError<SyncAllError>> =
         syncAllParser.tryParse(
             if (syncModel != null) {
-                syncAll(syncAllParser.encodeToString(config), syncModel)
+                app.lockbook.core.syncAll(syncModel)
             } else {
-                backgroundSync(syncAllParser.encodeToString(config))
+                app.lockbook.core.backgroundSync()
             }
         )
 
-    val writeToDocumentParser = Json {
+    private val writeToDocumentParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Unit.serializer(), WriteToDocumentError.serializer())
         }
     }
 
     fun writeToDocument(
-        config: Config,
         id: String,
         content: String
     ): Result<Unit, CoreError<WriteToDocumentError>> =
         writeToDocumentParser.tryParse(
-            writeDocument(
-                writeToDocumentParser.encodeToString(config),
+            app.lockbook.core.writeDocument(
                 id,
                 content
             )
         )
 
-    val getRootParser = Json {
+    private val getRootParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(DecryptedFileMetadata.serializer(), GetRootError.serializer())
         }
     }
 
-    fun getRoot(config: Config): Result<DecryptedFileMetadata, CoreError<GetRootError>> =
+    fun getRoot(): Result<DecryptedFileMetadata, CoreError<GetRootError>> =
         getRootParser.tryParse(
-            getRoot(
-                getRootParser.encodeToString(config)
-            )
+            app.lockbook.core.getRoot()
         )
 
-    val getAccountParser = Json {
+    private val getAccountParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Account.serializer(), GetAccountError.serializer())
         }
     }
 
-    fun getAccount(config: Config): Result<Account, CoreError<GetAccountError>> =
+    fun getAccount(): Result<Account, CoreError<GetAccountError>> =
         getAccountParser.tryParse(
-            getAccount(
-                getAccountParser.encodeToString(config)
-            )
+            app.lockbook.core.getAccount()
         )
 
     fun convertToHumanDuration(
         metadataVersion: Long
     ): String = app.lockbook.core.convertToHumanDuration(metadataVersion)
 
-    val getUsageParser = Json {
+    private val getUsageParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(UsageMetrics.serializer(), GetUsageError.serializer())
         }
     }
 
-    fun getUsage(
-        config: Config
-    ): Result<UsageMetrics, CoreError<GetUsageError>> =
+    fun getUsage(): Result<UsageMetrics, CoreError<GetUsageError>> =
         getUsageParser.tryParse(
-            getUsage(
-                getUsageParser.encodeToString(config)
-            )
+            app.lockbook.core.getUsage()
         )
 
-    val getUncompressedUsageParser = Json {
+    private val getUncompressedUsageParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(UsageItemMetric.serializer(), GetUsageError.serializer())
         }
     }
 
-    fun getUncompressedUsage(
-        config: Config
-    ): Result<UsageItemMetric, CoreError<GetUsageError>> =
+    fun getUncompressedUsage(): Result<UsageItemMetric, CoreError<GetUsageError>> =
         getUncompressedUsageParser.tryParse(
-            getUncompressedUsage(getUncompressedUsageParser.encodeToString(config))
+            app.lockbook.core.getUncompressedUsage()
         )
 
-    val getChildrenParser = Json {
+    private val getChildrenParser = Json {
         serializersModule = SerializersModule {
-            createPolyRelation(ListSerializer(DecryptedFileMetadata.serializer()), GetChildrenError.serializer())
+            createPolyRelation(
+                ListSerializer(DecryptedFileMetadata.serializer()),
+                GetChildrenError.serializer()
+            )
         }
     }
 
-    fun getChildren(
-        config: Config,
-        parentId: String
-    ): Result<List<DecryptedFileMetadata>, CoreError<GetChildrenError>> =
+    fun getChildren(parentId: String): Result<List<DecryptedFileMetadata>, CoreError<GetChildrenError>> =
         getChildrenParser.tryParse(
-            getChildren(getChildrenParser.encodeToString(config), parentId)
+            app.lockbook.core.getChildren(parentId)
         )
 
-    val getFileByIdParser = Json {
+    private val getFileByIdParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(DecryptedFileMetadata.serializer(), GetFileByIdError.serializer())
         }
     }
 
-    fun getFileById(
-        config: Config,
-        id: String
-    ): Result<DecryptedFileMetadata, CoreError<GetFileByIdError>> =
+    fun getFileById(id: String): Result<DecryptedFileMetadata, CoreError<GetFileByIdError>> =
         getFileByIdParser.tryParse(
-            getFileById(getFileByIdParser.encodeToString(config), id)
+            app.lockbook.core.getFileById(id)
         )
 
-    val readDocumentParser = Json {
+    private val readDocumentParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(String.serializer(), ReadDocumentError.serializer())
         }
     }
 
     fun readDocument(
-        config: Config,
         id: String
     ): Result<String, CoreError<ReadDocumentError>> =
         readDocumentParser.tryParse(
-            readDocument(
-                readDocumentParser.encodeToString(config),
-                id
-            )
+            app.lockbook.core.readDocument(id)
         )
 
-    val saveDocumentToDiskParser = Json {
+    private val saveDocumentToDiskParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Unit.serializer(), SaveDocumentToDiskError.serializer())
         }
     }
 
     fun saveDocumentToDisk(
-        config: Config,
         id: String,
         location: String
     ): Result<Unit, CoreError<SaveDocumentToDiskError>> =
         saveDocumentToDiskParser.tryParse(
-            saveDocumentToDisk(saveDocumentToDiskParser.encodeToString(config), id, location)
+            app.lockbook.core.saveDocumentToDisk(id, location)
         )
 
-    val exportDrawingToDiskParser = Json {
+    private val exportDrawingToDiskParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Unit.serializer(), ExportDrawingToDiskError.serializer())
         }
     }
 
     fun exportDrawingToDisk(
-        config: Config,
         id: String,
         format: SupportedImageFormats,
         location: String
     ): Result<Unit, CoreError<ExportDrawingToDiskError>> =
         exportDrawingToDiskParser.tryParse(
-            exportDrawingToDisk(
-                exportDrawingToDiskParser.encodeToString(config),
+            app.lockbook.core.exportDrawingToDisk(
                 id,
                 exportDrawingToDiskParser.encodeToString(format),
                 location
             )
         )
 
-    val createFileParser = Json {
+    private val createFileParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(DecryptedFileMetadata.serializer(), CreateFileError.serializer())
         }
     }
 
     fun createFile(
-        config: Config,
         parentId: String,
         name: String,
         fileType: FileType
     ): Result<DecryptedFileMetadata, CoreError<CreateFileError>> =
         createFileParser.tryParse(
-            createFile(
-                createFileParser.encodeToString(config),
+            app.lockbook.core.createFile(
                 name,
                 parentId,
                 createFileParser.encodeToString(fileType)
             )
         )
 
-    val deleteFileParser = Json {
+    private val deleteFileParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Unit.serializer(), FileDeleteError.serializer())
         }
     }
 
     fun deleteFile(
-        config: Config,
         id: String
     ): Result<Unit, CoreError<FileDeleteError>> =
         deleteFileParser.tryParse(
-            deleteFile(
-                deleteFileParser.encodeToString(
-                    config
-                ),
-                id
-            )
+            app.lockbook.core.deleteFile(id)
         )
 
-    val renameFileParser = Json {
+    private val renameFileParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Unit.serializer(), RenameFileError.serializer())
         }
     }
 
     fun renameFile(
-        config: Config,
         id: String,
         name: String
     ): Result<Unit, CoreError<RenameFileError>> =
         renameFileParser.tryParse(
-            renameFile(
-                renameFileParser.encodeToString(
-                    config
-                ),
-                id, name
-            )
+            app.lockbook.core.renameFile(id, name)
         )
 
-    val moveFileParser = Json {
+    private val moveFileParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(Unit.serializer(), MoveFileError.serializer())
         }
     }
 
     fun moveFile(
-        config: Config,
         id: String,
         parentId: String
     ): Result<Unit, CoreError<MoveFileError>> =
         moveFileParser.tryParse(
-            moveFile(
-                moveFileParser.encodeToString(
-                    config
-                ),
-                id, parentId
-            )
+            app.lockbook.core.moveFile(id, parentId)
         )
 
-    val calculateWorkParser = Json {
+    private val calculateWorkParser = Json {
         serializersModule = SerializersModule {
             createPolyRelation(WorkCalculated.serializer(), CalculateWorkError.serializer())
         }
     }
 
-    fun calculateWork(config: Config): Result<WorkCalculated, CoreError<CalculateWorkError>> =
+    fun calculateWork(): Result<WorkCalculated, CoreError<CalculateWorkError>> =
         calculateWorkParser.tryParse(
-            calculateWork(calculateWorkParser.encodeToString(config))
+            app.lockbook.core.calculateWork()
         )
 }