--- a/clients/cli/Cargo.toml
+++ b/clients/cli/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lockbook-cli"
-version = "0.3.17"
+version = "0.4.0"
 edition = "2018"
 
 [[bin]]