… Cargo.lock
… Cargo.toml
… clients/android/app/release/output-metadata.json
… clients/android/app/src/main/java/app/lockbook/model/SyncModel.kt
… clients/android/app/src/main/java/app/lockbook/model/TextEditorViewModel.kt
… clients/android/app/src/main/java/app/lockbook/screen/InitialLaunchFigureOuter.kt
… clients/android/app/src/main/java/app/lockbook/screen/OnBoardingActivity.kt
… clients/android/app/src/main/java/app/lockbook/screen/SettingsFragment.kt
… clients/android/app/src/main/java/app/lockbook/ui/CreateFileDialogFragment.kt
… clients/android/app/src/main/java/app/lockbook/ui/RenameFileDialogFragment.kt
… clients/android/app/src/main/java/app/lockbook/ui/UsageBarPreference.kt
… clients/android/app/src/main/java/app/lockbook/util/CoreData.kt
… clients/android/app/src/main/java/app/lockbook/util/Errors.kt
… clients/android/app/src/main/res/values/strings.xml
… clients/android/app/src/test/java/app/lockbook/AllErrorVariantsTest.kt
… clients/android/app/src/test/java/app/lockbook/CalculateWorkTest.kt
… clients/android/app/src/test/java/app/lockbook/CreateAccountTest.kt
… clients/android/app/src/test/java/app/lockbook/CreateFileTest.kt
… clients/android/app/src/test/java/app/lockbook/DeleteFileTest.kt
… clients/android/app/src/test/java/app/lockbook/ExportAccountTest.kt
… clients/android/app/src/test/java/app/lockbook/ExportDrawingToDiskTest.kt
… clients/android/app/src/test/java/app/lockbook/GetAccountTest.kt
… clients/android/app/src/test/java/app/lockbook/GetChildrenTest.kt
… clients/android/app/src/test/java/app/lockbook/GetDBStateTest.kt
… clients/android/app/src/test/java/app/lockbook/GetFileByIdTest.kt
… clients/android/app/src/test/java/app/lockbook/GetRootTest.kt
… clients/android/app/src/test/java/app/lockbook/GetUncompressedUsageTest.kt
… clients/android/app/src/test/java/app/lockbook/GetUsageTest.kt
… clients/android/app/src/test/java/app/lockbook/ImportAccountTest.kt
… clients/android/app/src/test/java/app/lockbook/InitLoggerTest.kt