# Notable Changes

- Core is now stateful, clients need to call `init()` prior to using it
- Config is no longer required for each call into core
- Migrations are no longer something clients need to explicitly model
- all test utils are moved to their own crate and are only a `dev-dependency` 
- reqwest client is put in a static location
- versions for everything are bumped to 0.4.0

+ fixes #869 
+ fixes #961 

## Work tracking

- [x] android
- [x] windows
- [x] apple
- [x] atomic document writes
- [x] restore logs
- [x] core
- [x] tests
- [x] linux
- [x] cli

## Benchmarks

Benchmarks across the board are very promising. The few places where we see regressions to "app opening time" but the old benchmarks aren't representative of what's going on. For instance any application needs to, on open, check db state, check if there's an account then list the files. Hmdb will do all these steps in one go (as quickly as possible) on startup anyways. So the benchmarks penalize it. But even so after more than 50 files, hmdb pulls away. I was thinking about fixing these benchmarks in the old version, but the old core takes 9.5 hours to complete benchmarks. So, basically, it just is strictly faster in every situation.

A general summary of the situation:

![image](https://user-images.githubusercontent.com/821972/164772440-9d852f32-9b64-42bf-bae5-cabfde5544de.png)

```
create_file/1           time:   [15.253 ms 18.376 ms 20.274 ms]                         
                        thrpt:  [49.325  elem/s 54.417  elem/s 65.562  elem/s]
                 change:
                        time:   [-54.776% -45.628% -34.016%] (p = 0.00 < 0.05)
                        thrpt:  [+51.553% +83.920% +121.12%]
                        Performance has improved.
create_file/10          time:   [197.21 ms 246.14 ms 276.32 ms]                          
                        thrpt:  [36.190  elem/s 40.627  elem/s 50.707  elem/s]
                 change:
                        time:   [-51.994% -41.278% -28.079%] (p = 0.00 < 0.05)
                        thrpt:  [+39.041% +70.293% +108.31%]
                        Performance has improved.
create_file/100         time:   [1.6368 s 2.1695 s 2.7209 s]                             
                        thrpt:  [36.753  elem/s 46.093  elem/s 61.094  elem/s]
                 change:
                        time:   [-69.587% -58.217% -45.284%] (p = 0.00 < 0.05)
                        thrpt:  [+82.762% +139.33% +228.80%]
                        Performance has improved.
create_file/500         time:   [24.247 s 34.397 s 44.890 s]                             
                        thrpt:  [11.138  elem/s 14.536  elem/s 20.621  elem/s]
                 change:
                        time:   [-70.564% -53.835% -25.614%] (p = 0.01 < 0.05)
                        thrpt:  [+34.434% +116.61% +239.72%]
                        Performance has improved.
create_file/1000        time:   [82.285 s 140.86 s 204.23 s]                              
                        thrpt:  [4.8964  elem/s 7.0991  elem/s 12.153  elem/s]
                 change:
                        time:   [-77.745% -58.542% -28.532%] (p = 0.01 < 0.05)
                        thrpt:  [+39.923% +141.21% +349.33%]
                        Performance has improved.
create_file/2000        time:   [591.08 s 1132.2 s 1725.4 s]                              
                        thrpt:  [1.1592  elem/s 1.7665  elem/s 3.3836  elem/s]
                 change:
                        time:   [-69.820% -40.661% +14.642%] (p = 0.15 > 0.05)
                        thrpt:  [-12.772% +68.523% +231.34%]
                        No change in performance detected.
open_app_get_state      time:   [4.7479 us 4.7595 us 4.7748 us]                               
                        change: [+45.972% +47.684% +49.273%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe

open_app_get_account    time:   [593.15 us 750.35 us 842.39 us]                                
                        change: [+26110% +30057% +35499%] (p = 0.00 < 0.05)
                        Performance has regressed.
open_app_list_metadatas/1                                                                           
                        time:   [643.62 us 742.63 us 800.42 us]
                        change: [+325.93% +371.83% +414.61%] (p = 0.00 < 0.05)
                        Performance has regressed.
open_app_list_metadatas/10                                                                           
                        time:   [659.86 us 752.53 us 807.64 us]
                        change: [+201.77% +230.85% +259.00%] (p = 0.00 < 0.05)
                        Performance has regressed.
open_app_list_metadatas/100                                                                           
                        time:   [661.07 us 731.86 us 773.32 us]
                        change: [-29.467% -24.772% -18.983%] (p = 0.00 < 0.05)
                        Performance has improved.
open_app_list_metadatas/500                                                                            
                        time:   [1.4840 ms 1.5080 ms 1.5217 ms]
                        change: [-62.899% -62.528% -62.126%] (p = 0.00 < 0.05)
                        Performance has improved.
open_app_list_metadatas/1000                                                                            
                        time:   [2.6639 ms 2.6800 ms 2.6889 ms]
                        change: [-66.343% -66.029% -65.792%] (p = 0.00 < 0.05)
                        Performance has improved.
open_app_list_metadatas/2000                                                                           
                        time:   [5.2975 ms 5.3096 ms 5.3193 ms]
                        change: [-67.992% -67.653% -67.352%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe

open_app_list_paths/1   time:   [653.36 us 752.33 us 810.03 us]                                 
                        change: [+326.86% +374.80% +420.89%] (p = 0.00 < 0.05)
                        Performance has regressed.
open_app_list_paths/10  time:   [671.91 us 764.60 us 818.87 us]                                  
                        change: [+201.58% +230.21% +257.28%] (p = 0.00 < 0.05)
                        Performance has regressed.
open_app_list_paths/100 time:   [679.96 us 747.90 us 788.10 us]                                   
                        change: [-30.560% -25.800% -20.371%] (p = 0.00 < 0.05)
                        Performance has improved.
open_app_list_paths/500 time:   [1.5563 ms 1.5797 ms 1.5936 ms]                                    
                        change: [-62.280% -61.874% -61.474%] (p = 0.00 < 0.05)
                        Performance has improved.
open_app_list_paths/1000                                                                            
                        time:   [2.9240 ms 2.9359 ms 2.9439 ms]
                        change: [-64.809% -64.234% -63.862%] (p = 0.00 < 0.05)
                        Performance has improved.
open_app_list_paths/2000                                                                           
                        time:   [5.7707 ms 5.7821 ms 5.7891 ms]
                        change: [-66.040% -65.765% -65.470%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe

sync/1                  time:   [13.963 ms 14.762 ms 15.305 ms]                  
                        thrpt:  [65.339  elem/s 67.743  elem/s 71.616  elem/s]
                 change:
                        time:   [-13.791% -8.0596% -3.5875%] (p = 0.01 < 0.05)
                        thrpt:  [+3.7210% +8.7661% +15.997%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) low mild
sync/10                 time:   [56.673 ms 58.134 ms 59.362 ms]                   
                        thrpt:  [168.46  elem/s 172.02  elem/s 176.45  elem/s]
                 change:
                        time:   [-19.634% -17.042% -14.678%] (p = 0.00 < 0.05)
                        thrpt:  [+17.204% +20.543% +24.430%]
                        Performance has improved.
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 6.7s.
sync/100                time:   [652.07 ms 668.27 ms 686.07 ms]                   
                        thrpt:  [145.76  elem/s 149.64  elem/s 153.36  elem/s]
                 change:
                        time:   [-52.070% -50.922% -49.773%] (p = 0.00 < 0.05)
                        thrpt:  [+99.095% +103.76% +108.64%]
                        Performance has improved.
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 73.8s.
sync/500                time:   [7.2198 s 7.2622 s 7.3081 s]                      
                        thrpt:  [68.417  elem/s 68.849  elem/s 69.254  elem/s]
                 change:
                        time:   [-69.650% -69.472% -69.272%] (p = 0.00 < 0.05)
                        thrpt:  [+225.44% +227.57% +229.49%]
                        Performance has improved.
Benchmarking sync/1000: Warming up for 3.0000 s
sync/1000               time:   [25.141 s 25.480 s 25.793 s]                       
                        thrpt:  [38.770  elem/s 39.247  elem/s 39.775  elem/s]
                 change:
                        time:   [-72.747% -72.361% -71.971%] (p = 0.00 < 0.05)
                        thrpt:  [+256.78% +261.80% +266.94%]
                        Performance has improved.
Benchmarking sync/2000: Warming up for 3.0000 s
sync/2000               time:   [99.747 s 101.81 s 103.60 s]                       
                        thrpt:  [19.306  elem/s 19.645  elem/s 20.051  elem/s]
                 change:
                        time:   [-73.553% -73.014% -72.479%] (p = 0.00 < 0.05)
                        thrpt:  [+263.36% +270.56% +278.11%]
                        Performance has improved.

test_repo_integrity/1   time:   [257.02 us 257.91 us 259.42 us]                                 
                        thrpt:  [3.8547 Kelem/s 3.8773 Kelem/s 3.8907 Kelem/s]
                 change:
                        time:   [-21.045% -20.544% -20.094%] (p = 0.00 < 0.05)
                        thrpt:  [+25.147% +25.856% +26.654%]
                        Performance has improved.
test_repo_integrity/10  time:   [1.6097 ms 1.6141 ms 1.6174 ms]                                   
                        thrpt:  [6.1826 Kelem/s 6.1953 Kelem/s 6.2122 Kelem/s]
                 change:
                        time:   [-36.126% -35.689% -35.110%] (p = 0.00 < 0.05)
                        thrpt:  [+54.107% +55.494% +56.559%]
                        Performance has improved.
Found 3 outliers among 10 measurements (30.00%)
  1 (10.00%) low mild
  1 (10.00%) high mild
  1 (10.00%) high severe
test_repo_integrity/100 time:   [28.981 ms 29.132 ms 29.215 ms]                                   
                        thrpt:  [3.4229 Kelem/s 3.4326 Kelem/s 3.4505 Kelem/s]
                 change:
                        time:   [-68.989% -68.847% -68.721%] (p = 0.00 < 0.05)
                        thrpt:  [+219.70% +220.99% +222.46%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) low severe
test_repo_integrity/500 time:   [460.53 ms 461.60 ms 462.72 ms]                                  
                        thrpt:  [1.0806 Kelem/s 1.0832 Kelem/s 1.0857 Kelem/s]
                 change:
                        time:   [-77.224% -77.108% -76.969%] (p = 0.00 < 0.05)
                        thrpt:  [+334.20% +336.83% +339.05%]
                        Performance has improved.
Benchmarking test_repo_integrity/1000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 16.9s.
test_repo_integrity/1000                                                                          
                        time:   [1.6790 s 1.6907 s 1.7021 s]
                        thrpt:  [587.51  elem/s 591.45  elem/s 595.60  elem/s]
                 change:
                        time:   [-79.111% -78.926% -78.712%] (p = 0.00 < 0.05)
                        thrpt:  [+369.75% +374.51% +378.73%]
                        Performance has improved.
Benchmarking test_repo_integrity/2000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 66.4s.
test_repo_integrity/2000                                                                          
                        time:   [6.5193 s 6.5543 s 6.5865 s]
                        thrpt:  [303.65  elem/s 305.14  elem/s 306.78  elem/s]
                 change:
                        time:   [-80.583% -80.464% -80.344%] (p = 0.00 < 0.05)
                        thrpt:  [+408.76% +411.89% +415.01%]
                        Performance has improved.
Found 3 outliers among 10 measurements (30.00%)
  2 (20.00%) low mild
  1 (10.00%) high mild

write_file/100          time:   [18.642 ms 21.480 ms 23.192 ms]                          
                        thrpt:  [4.3119 Kelem/s 4.6555 Kelem/s 5.3643 Kelem/s]
                 change:
                        time:   [-55.493% -48.914% -41.307%] (p = 0.00 < 0.05)
                        thrpt:  [+70.378% +95.750% +124.68%]
                        Performance has improved.
write_file/1000         time:   [18.495 ms 21.464 ms 23.255 ms]                           
                        thrpt:  [43.002 Kelem/s 46.590 Kelem/s 54.070 Kelem/s]
                 change:
                        time:   [-55.583% -49.261% -41.860%] (p = 0.00 < 0.05)
                        thrpt:  [+72.000% +97.086% +125.14%]
                        Performance has improved.
write_file/2000         time:   [18.880 ms 21.670 ms 23.319 ms]                           
                        thrpt:  [85.768 Kelem/s 92.295 Kelem/s 105.93 Kelem/s]
                 change:
                        time:   [-55.169% -48.682% -41.234%] (p = 0.00 < 0.05)
                        thrpt:  [+70.166% +94.864% +123.06%]
                        Performance has improved.
write_file/5000         time:   [18.916 ms 21.518 ms 23.031 ms]                           
                        thrpt:  [217.10 Kelem/s 232.37 Kelem/s 264.33 Kelem/s]
                 change:
                        time:   [-55.208% -48.607% -41.226%] (p = 0.00 < 0.05)
                        thrpt:  [+70.142% +94.579% +123.26%]
                        Performance has improved.
write_file/10000        time:   [18.840 ms 21.688 ms 23.405 ms]                            
                        thrpt:  [427.26 Kelem/s 461.08 Kelem/s 530.79 Kelem/s]
                 change:
                        time:   [-54.081% -48.001% -40.776%] (p = 0.00 < 0.05)
                        thrpt:  [+68.850% +92.312% +117.78%]
                        Performance has improved.
write_file/100000       time:   [13.810 ms 16.478 ms 18.087 ms]                             
                        thrpt:  [5.5290 Melem/s 6.0688 Melem/s 7.2414 Melem/s]
                 change:
                        time:   [-58.200% -48.716% -36.841%] (p = 0.00 < 0.05)
                        thrpt:  [+58.330% +94.992% +139.23%]
                        Performance has improved.


________________________________________________________
Executed in  262.75 mins    fish           external
   usr time  262.58 mins    0.00 micros  262.58 mins
   sys time    4.44 mins  360.00 micros    4.44 mins
```

[before.zip](https://github.com/lockbook/lockbook/files/8522189/before.zip)
[after.zip](https://github.com/lockbook/lockbook/files/8543058/after.zip)