@steverusso 2022-04-20
### Tested CLI
Debug build for commit https://github.com/lockbook/lockbook/pull/1053/commits/474cbded6c7c6658daeda507612564d857aaa279, on Manjaro 21.2.6, the following commands were tested against a local server setup:
- [x] errors
- [x] help
- [x] new-account
- [x] whoami
- [x] new
- [x] print
- [x] edit
- [x] list, list-all, list-docs, list-folders
- [x] status
- [x] sync
- [x] rename
- [x] move
- [x] remove
- [x] get-usage
- [x] tree
- [x] validate
- [x] backup
- [x] copy
- [x] export-private-key
- [x] import-private-key

Not tested:
- [ ] export-drawing