@steverusso 2022-04-21
### Testing Linux

Debug build for https://github.com/lockbook/lockbook/pull/1053/commits/474cbded6c7c6658daeda507612564d857aaa279, Manjaro 21.2.6, performed the following against a local server setup:
- [x] Created a new account
- [x] Created a new document and folder
- [x] Added content to that file
- [x] Moved that file into the folder
- [x] Renamed the file
- [x] Deleted the file
- [x] Created new file in new sub directory of original folder
- [x] Searched files
- [x] Exported account string
- [x] Imported account

Auto syncing and auto saving the whole time. Threw some manual saves and syncs in there. Usage status in side panel and usage data in settings both good.